Migrate to GH actions
This is currently of questionable usage, as the test suite is
just the templated out version. Still, it does more than the
travis tests, which literally just installed bundler.

However, the entire gem is pretty much just configuration
and boiler-plate, so there isn't a huge amount to test. Plus
with the planned migration to the V2 api, the gem will either be
dropped, or significantly re-worked shortly.